Remove inheritance
