+++ DuckDuckGo/TabViewController.swift
@@ -23,30 +23,96 @@ import Core
 import Device
 import StoreKit
 
-class TabViewController: WebViewController {
+class TabViewController: UIViewController {
+    
+    private struct Constants {
+        static let unsupportedUrlErrorCode = -1002
+        static let urlCouldNotBeLoaded = 101
+        static let frameLoadInterruptedErrorCode = 102
+        static let minimumProgress: Float = 0.1
 
-    struct Constants {
         // swiftlint:disable line_length
         static let desktopUserAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/12.0 Safari/605.1.15"
         // swiftlint:enable line_length
     }
     
+    @IBOutlet weak var progressBar: UIProgressView!
+    @IBOutlet weak var error: UIView!
+    @IBOutlet weak var errorMessage: UILabel!
+    @IBOutlet weak var webViewContainer: UIView!
     @IBOutlet var showBarsTapGestureRecogniser: UITapGestureRecognizer!
 
     weak var delegate: TabDelegate?
     weak var chromeDelegate: BrowserChromeDelegate?
 
-    private lazy var appUrls: AppUrls = AppUrls()
+    private(set) var webView: WKWebView!
     private lazy var appRatingPrompt: AppRatingPrompt = AppRatingPrompt()
-    private lazy var disconnectMeStore = DisconnectMeStore()
-    
-    private(set) var contentBlocker: ContentBlockerConfigurationStore!
     private weak var privacyController: PrivacyProtectionController?
+    
+    private lazy var appUrls: AppUrls = AppUrls()
+    private lazy var tld = TLD()
+    private lazy var statisticsStore: StatisticsStore = StatisticsUserDefaults()
+    private lazy var disconnectMeStore = DisconnectMeStore()
+    private var contentBlocker: ContentBlockerConfigurationStore!
+    private var httpsUpgrade = HTTPSUpgrade.shared
+
     private(set) var siteRating: SiteRating?
     private(set) var tabModel: Tab
-    
     private var httpsForced: Bool = false
+    private var lastUpgradedDomain: String?
+    private var lastError: Error?
+    private var shouldReloadOnError = false
+    private var failingUrls = Set<String>()
+    private var tearDownCount = 0
     
+    public var url: URL? {
+        didSet {
+            delegate?.tabLoadingStateDidChange(tab: self)
+        }
+    }
+    
+    public var name: String? {
+        return webView.title
+    }
+    
+    public var canGoBack: Bool {
+        let webViewCanGoBack = webView.canGoBack
+        let navigatedToError = webView.url != nil && isError
+        return webViewCanGoBack || navigatedToError
+    }
+    
+    public var canGoForward: Bool {
+        let webViewCanGoForward = webView.canGoForward
+        return webViewCanGoForward && !isError
+    }
+    
+    public var isError: Bool {
+        return !error.isHidden
+    }
+    
+    public var errorText: String? {
+        return errorMessage.text
+    }
+    
+    public var link: Link? {
+        if isError {
+            if let url = url ?? webView.url ?? URL(string: "") {
+                return Link(title: errorText, url: url)
+            }
+        }
+        
+        guard let url = url else {
+            return tabModel.link
+        }
+        
+        let activeLink = Link(title: name, url: url)
+        guard let storedLink = tabModel.link else {
+            return activeLink
+        }
+        
+        return activeLink.merge(with: storedLink)
+    }
+
     static func loadFromStoryboard(model: Tab, contentBlocker: ContentBlockerConfigurationStore) -> TabViewController {
         let storyboard = UIStoryboard(name: "Tab", bundle: nil)
         guard let controller = storyboard.instantiateViewController(withIdentifier: "TabViewController") as? TabViewController else {
@@ -60,28 +126,8 @@ class TabViewController: WebViewController {
     required init?(coder aDecoder: NSCoder) {
         tabModel = Tab(link: nil)
         super.init(coder: aDecoder)
-        webEventsDelegate = self
     }
-
-    public var link: Link? {
-        if isError {
-            if let url = loadedURL ?? webView.url ?? URL(string: "") {
-                return Link(title: errorText, url: url)
-            }
-        }
-
-        guard let url = url else {
-            return tabModel.link
-        }
-
-        let activeLink = Link(title: name, url: url)
-        guard let storedLink = tabModel.link else {
-            return activeLink
-        }
-
-        return activeLink.merge(with: storedLink)
-    }
-
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         addContentBlockerConfigurationObserver()
@@ -90,6 +136,203 @@ class TabViewController: WebViewController {
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         resetNavigationBar()
+    }
+    
+    @objc func onApplicationWillResignActive() {
+        shouldReloadOnError = true
+    }
+    
+    func attachWebView(configuration: WKWebViewConfiguration, andLoadUrl url: URL?, consumeCookies: Bool) {
+        webView = WKWebView(frame: view.bounds, configuration: configuration)
+        webView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        attachLongPressHandler(webView: webView)
+        webView.allowsBackForwardNavigationGestures = true
+        
+        addObservers()
+        
+        webView.navigationDelegate = self
+        webView.uiDelegate = self
+        webViewContainer.addSubview(webView)
+        webView.configuration.userContentController.add(self, name: MessageHandlerNames.trackerDetected)
+        webView.configuration.userContentController.add(self, name: MessageHandlerNames.cache)
+        webView.configuration.userContentController.add(self, name: MessageHandlerNames.log)
+        reloadScripts()
+        updateUserAgent()
+
+        if consumeCookies {
+            consumeCookiesThenLoadUrl(url)
+        } else if let url = url {
+            load(url: url)
+        }
+    }
+    
+    private func addObservers() {
+        webView.addObserver(self, forKeyPath: #keyPath(WKWebView.estimatedProgress), options: .new, context: nil)
+        webView.addObserver(self, forKeyPath: #keyPath(WKWebView.hasOnlySecureContent), options: .new, context: nil)
+        webView.addObserver(self, forKeyPath: #keyPath(WKWebView.url), options: .new, context: nil)
+        webView.addObserver(self, forKeyPath: #keyPath(WKWebView.canGoBack), options: .new, context: nil)
+        webView.addObserver(self, forKeyPath: #keyPath(WKWebView.canGoForward), options: .new, context: nil)
+    }
+    
+    private func attachLongPressHandler(webView: WKWebView) {
+        let handler = WebLongPressGestureRecognizer(target: self, action: #selector(onLongPress(sender:)))
+        handler.delegate = self
+        webView.scrollView.addGestureRecognizer(handler)
+    }
+    
+    private func consumeCookiesThenLoadUrl(_ url: URL?) {
+        webView.configuration.websiteDataStore.fetchDataRecords(ofTypes: WKWebsiteDataStore.allWebsiteDataTypes()) { _ in
+            WebCacheManager.consumeCookies()
+            if let url = url {
+                self.load(url: url)
+            }
+        }
+        
+        if url != nil {
+            progressBar.progress = Constants.minimumProgress
+            delegate?.tabLoadingStateDidChange(tab: self)
+            onWebpageDidStartLoading(httpsForced: false)
+        }
+    }
+    
+    public func load(url: URL) {
+        self.url = url
+        lastError = nil
+        updateUserAgent()
+        load(urlRequest: URLRequest(url: url))
+    }
+    
+    private func load(urlRequest: URLRequest) {
+        loadViewIfNeeded()
+        webView.stopLoading()
+        webView.load(urlRequest)
+    }
+    
+    open override func observeValue(forKeyPath keyPath: String?,
+                                    of object: Any?,
+                                    change: [NSKeyValueChangeKey: Any]?,
+                                    context: UnsafeMutableRawPointer?) {
+        
+        guard let keyPath = keyPath else { return }
+        
+        switch keyPath {
+            
+        case WebViewKeyPaths.estimatedProgress:
+            progressBar.progress = max(Constants.minimumProgress, Float(webView.estimatedProgress))
+            
+        case WebViewKeyPaths.hasOnlySecureContent:
+            hasOnlySecureContentChanged(hasOnlySecureContent: webView.hasOnlySecureContent)
+            
+        case WebViewKeyPaths.url:
+            urlDidChange()
+            
+        case WebViewKeyPaths.canGoBack:
+            delegate?.tabLoadingStateDidChange(tab: self)
+            
+        case WebViewKeyPaths.canGoForward:
+            delegate?.tabLoadingStateDidChange(tab: self)
+            
+        default:
+            Logger.log(text: "Unhandled keyPath \(keyPath)")
+        }
+    }
+    
+    func hasOnlySecureContentChanged(hasOnlySecureContent: Bool) {
+        guard webView.url?.host == siteRating?.url.host else { return }
+        siteRating?.hasOnlySecureContent = hasOnlySecureContent
+        updateSiteRating()
+    }
+    
+    private func urlDidChange() {
+        if self.url?.host == self.webView.url?.host {
+            self.url = self.webView.url
+        }
+    }
+    
+    private func checkForReloadOnError() {
+        guard shouldReloadOnError else { return }
+        shouldReloadOnError = false
+        reload()
+    }
+    
+    private func shouldReissueSearch(for url: URL) -> Bool {
+        return appUrls.isDuckDuckGoSearch(url: url) && !appUrls.hasCorrectMobileStatsParams(url: url)
+    }
+    
+    private func reissueSearchWithStatsParams(for url: URL) {
+        let mobileSearch = appUrls.applyStatsParams(for: url)
+        load(url: mobileSearch)
+    }
+    
+    private func showProgressIndicator() {
+        progressBar.alpha = 1
+        progressBar.progress = Constants.minimumProgress
+    }
+    
+    private func hideProgressIndicator() {
+        UIView.animate(withDuration: 1) {
+            self.progressBar.alpha = 0
+        }
+    }
+    
+    public func reload() {
+        updateUserAgent()
+        webView.reload()
+    }
+    
+    private func updateUserAgent() {
+        let userAgent = tabModel.isDesktop ? Constants.desktopUserAgent : nil
+        webView.customUserAgent = userAgent
+    }
+    
+    func goBack() {
+        if isError {
+            hideErrorMessage()
+            url = webView.url
+            onWebpageDidStartLoading(httpsForced: false)
+            onWebpageDidFinishLoading()
+        } else {
+            webView.goBack()
+        }
+    }
+    
+    func goForward() {
+        webView.goForward()
+    }
+    
+    @objc func onLongPress(sender: UILongPressGestureRecognizer) {
+        guard sender.state == .began else { return }
+        
+        let x = Int(sender.location(in: webView).x)
+        let y = Int(sender.location(in: webView).y)
+        let offsetY = y
+        
+        webView.getUrlAtPoint(x: x, y: offsetY) { [weak self] (url) in
+            guard let url = url else { return }
+            let point = Point(x: x, y: y)
+            self?.launchLongPressMenu(atPoint: point, forUrl: url)
+        }
+    }
+    
+    private func showError(message: String) {
+        webView.isHidden = true
+        error.isHidden = false
+        errorMessage.text = message
+    }
+    
+    private func hideErrorMessage() {
+        error.isHidden = true
+        webView.isHidden = false
+    }
+    
+    private func reloadScripts() {
+        webView.configuration.userContentController.removeAllUserScripts()
+        webView.configuration.loadScripts(contentBlocking: !isDuckDuckGoUrl())
+    }
+    
+    private func isDuckDuckGoUrl() -> Bool {
+        guard let url = url else { return false }
+        return appUrls.isDuckDuckGo(url: url)
     }
 
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
@@ -111,7 +354,6 @@ class TabViewController: WebViewController {
             controller.siteRating = siteRating
             controller.errorText = isError ? errorText : nil
         }
-
     }
 
     private func addContentBlockerConfigurationObserver() {
@@ -122,7 +364,7 @@ class TabViewController: WebViewController {
     }
 
     @objc func onContentBlockerConfigurationChanged() {
-        // defer it for 0.2s so that the privacy protection UI can update instantly, otherwise this causes a visible delay
+        // defer for 0.2s so that the privacy protection UI can update without a visible delay
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
             self.webView?.reload()
         }
@@ -130,8 +372,7 @@ class TabViewController: WebViewController {
 
     func webView(_ webView: WKWebView, didReceiveServerRedirectForProvisionalNavigation navigation: WKNavigation!) {
         guard let url = webView.url else { return }
-
-        loadedURL = url
+        self.url = url
         self.siteRating = SiteRating(url: url, httpsForced: httpsForced)
         updateSiteRating()
     }
@@ -144,7 +385,7 @@ class TabViewController: WebViewController {
         showBars(animated: false)
     }
 
-    fileprivate func showBars(animated: Bool = true) {
+    private func showBars(animated: Bool = true) {
         chromeDelegate?.setBarsHidden(false, animated: animated)
     }
 
@@ -157,7 +398,7 @@ class TabViewController: WebViewController {
         }
     }
 
-    fileprivate func resetSiteRating() {
+    private func resetSiteRating() {
         if let url = url {
             siteRating = SiteRating(url: url, httpsForced: httpsForced)
         } else {
@@ -166,14 +407,14 @@ class TabViewController: WebViewController {
         onSiteRatingChanged()
     }
 
-    fileprivate func updateSiteRating() {
+    private func updateSiteRating() {
         if isError {
             siteRating = nil
         }
         onSiteRatingChanged()
     }
 
-    fileprivate func onSiteRatingChanged() {
+    private func onSiteRatingChanged() {
         delegate?.tab(self, didChangeSiteRating: siteRating)
         privacyController?.updateSiteRating(siteRating)
     }
@@ -373,7 +614,7 @@ class TabViewController: WebViewController {
         }
     }
 
-    fileprivate func shouldLoad(url: URL, forDocument documentUrl: URL) -> Bool {
+    private func shouldLoad(url: URL, forDocument documentUrl: URL) -> Bool {
         if shouldOpenExternally(url: url) {
             openExternally(url: url)
             return false
@@ -423,6 +664,26 @@ class TabViewController: WebViewController {
         removeFromParent()
         view.removeFromSuperview()
     }
+    
+    public func tearDown() {
+        guard tearDownCount == 0 else {
+            fatalError("tearDown has already happened")
+        }
+        tearDownCount += 1
+        removeObservers()
+        webView.removeFromSuperview()
+        webView.configuration.userContentController.removeScriptMessageHandler(forName: MessageHandlerNames.trackerDetected)
+        webView.configuration.userContentController.removeScriptMessageHandler(forName: MessageHandlerNames.cache)
+        webView.configuration.userContentController.removeScriptMessageHandler(forName: MessageHandlerNames.log)
+    }
+    
+    private func removeObservers() {
+        webView.removeObserver(self, forKeyPath: #keyPath(WKWebView.estimatedProgress))
+        webView.removeObserver(self, forKeyPath: #keyPath(WKWebView.hasOnlySecureContent))
+        webView.removeObserver(self, forKeyPath: #keyPath(WKWebView.url))
+        webView.removeObserver(self, forKeyPath: #keyPath(WKWebView.canGoForward))
+        webView.removeObserver(self, forKeyPath: #keyPath(WKWebView.canGoBack))
+    }
 
     func destroy() {
         dismiss()
@@ -430,13 +691,13 @@ class TabViewController: WebViewController {
     }
 }
 
-private struct MessageHandlerNames {
-    static let trackerDetected = "trackerDetectedMessage"
-    static let cache = "cacheMessage"
-    static let log = "log"
-}
-
 extension TabViewController: WKScriptMessageHandler {
+    
+    private struct MessageHandlerNames {
+        static let trackerDetected = "trackerDetectedMessage"
+        static let cache = "cacheMessage"
+        static let log = "log"
+    }
 
     public func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
 
@@ -453,9 +714,7 @@ extension TabViewController: WKScriptMessageHandler {
 
         default:
             assertionFailure("Unhandled message: \(message.name)")
-
         }
-
     }
 
     private func handleLog(message: WKScriptMessage) {
@@ -510,50 +769,47 @@ extension TabViewController: WKScriptMessageHandler {
     }
 }
 
-extension TabViewController: WebEventsDelegate {
-
-    func attached(webView: WKWebView) {
-        webView.configuration.userContentController.add(self, name: MessageHandlerNames.trackerDetected)
-        webView.configuration.userContentController.add(self, name: MessageHandlerNames.cache)
-        webView.configuration.userContentController.add(self, name: MessageHandlerNames.log)
-        reloadScripts()
-        updateUserAgent()
+extension TabViewController: WKNavigationDelegate {
+    
+    public func webView(_ webView: WKWebView,
+                        didReceive challenge: URLAuthenticationChallenge,
+                        completionHandler: @escaping (URLSession.AuthChallengeDisposition, URLCredential?) -> Void) {
+        
+        completionHandler(.performDefaultHandling, nil)
+        guard let serverTrust = challenge.protectionSpace.serverTrust else { return }
+        ServerTrustCache.shared.put(serverTrust: serverTrust, forDomain: challenge.protectionSpace.host)
     }
     
-    private func updateUserAgent() {
-        userAgent = tabModel.isDesktop ? Constants.desktopUserAgent : nil
+    public func webView(_ webView: WKWebView, didCommit navigation: WKNavigation!) {
+        url = webView.url
+        let httpsForced = tld.domain(lastUpgradedDomain) == tld.domain(webView.url?.host)
+        onWebpageDidStartLoading(httpsForced: httpsForced)
+        
+        if let url = webView.url, isHttpsUpgradeSite(url: url) {
+            statisticsStore.httpsUpgradesTotal += 1
+        }
     }
-
-    func detached(webView: WKWebView) {
-        webView.configuration.userContentController.removeScriptMessageHandler(forName: MessageHandlerNames.trackerDetected)
-        webView.configuration.userContentController.removeScriptMessageHandler(forName: MessageHandlerNames.cache)
-        webView.configuration.userContentController.removeScriptMessageHandler(forName: MessageHandlerNames.log)
-    }
-
-    func contentProcessDidTerminate(webView: WKWebView) {
-        delegate?.tabContentProcessDidTerminate(tab: self)
-    }
-
-    func webpageDidStartLoading(httpsForced: Bool) {
+    
+    private func onWebpageDidStartLoading(httpsForced: Bool) {
         Logger.log(items: "webpageLoading started:", Date().timeIntervalSince1970)
         self.httpsForced = httpsForced
         delegate?.showBars()
-
-        // if host and scheme are the same, use same protection id and don't inject scripts, otherwise, reset and reload
+        
+        // if host and scheme are the same, don't inject scripts, otherwise, reset and reload
         if let siteRating = siteRating, siteRating.url.host == url?.host, siteRating.url.scheme == url?.scheme {
             self.siteRating = SiteRating(url: siteRating.url, httpsForced: httpsForced)
         } else {
             resetSiteRating()
         }
-
+        
         tabModel.link = link
         delegate?.tabLoadingStateDidChange(tab: self)
         UIApplication.shared.isNetworkActivityIndicatorVisible = true
-
+        
         if let domain = siteRating?.domain {
             NetworkLeaderboard.shared.visited(domain: domain)
         }
-
+        
         if #available(iOS 10.3, *) {
             appRatingPrompt.registerUsage()
             if appRatingPrompt.shouldPrompt() {
@@ -561,10 +817,26 @@ extension TabViewController: WebEventsDelegate {
                 appRatingPrompt.shown()
             }
         }
-        
     }
-
-    func webpageDidFinishLoading() {
+    
+    public func webView(_ webView: WKWebView, decidePolicyFor navigationResponse: WKNavigationResponse, decisionHandler: @escaping (WKNavigationResponsePolicy) -> Void) {
+        decisionHandler(.allow)
+        url = webView.url
+    }
+    
+    public func webView(_ webView: WKWebView, didStartProvisionalNavigation navigation: WKNavigation!) {
+        lastError = nil
+        shouldReloadOnError = false
+        hideErrorMessage()
+        showProgressIndicator()
+    }
+    
+    public func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
+        hideProgressIndicator()
+        onWebpageDidFinishLoading()
+    }
+    
+    private func onWebpageDidFinishLoading() {
         Logger.log(items: "webpageLoading finished:", Date().timeIntervalSince1970)
         siteRating?.finishedLoading = true
         updateSiteRating()
@@ -572,8 +844,20 @@ extension TabViewController: WebEventsDelegate {
         UIApplication.shared.isNetworkActivityIndicatorVisible = false
         delegate?.tabLoadingStateDidChange(tab: self)
     }
+    
+    public func webView(_ webView: WKWebView, didFail navigation: WKNavigation!, withError error: Error) {
+        hideProgressIndicator()
+        webpageDidFailToLoad()
+        
+        let error = error as NSError
+        if let url = webView.url, isHttpsUpgradeSite(url: url) {
+            reportHttpsUpgradeSiteError(url: url, error: "\(error.domain)_\(error.code)")
+        }
+        
+        checkForReloadOnError()
+    }
 
-    func webpageDidFailToLoad() {
+    private func webpageDidFailToLoad() {
         Logger.log(items: "webpageLoading failed:", Date().timeIntervalSince1970)
         if isError {
             showBars(animated: true)
@@ -583,29 +867,137 @@ extension TabViewController: WebEventsDelegate {
         UIApplication.shared.isNetworkActivityIndicatorVisible = false
         self.delegate?.tabLoadingStateDidChange(tab: self)
     }
-
-    func webView(_ webView: WKWebView, shouldLoadUrl url: URL, forDocument documentUrl: URL) -> Bool {
-        return shouldLoad(url: url, forDocument: documentUrl)
+    
+    public func webView(_ webView: WKWebView, didFailProvisionalNavigation navigation: WKNavigation!, withError error: Error) {
+        hideProgressIndicator()
+        lastError = error
+        let error = error as NSError
+        
+        if let url = url, isHttpsUpgradeSite(url: url) {
+            reportHttpsUpgradeSiteError(url: url, error: "\(error.domain)_\(error.code)")
+        }
+        
+        // prevent loops where a site keeps redirecting to itself (e.g. bbc)
+        if let url = url,
+            let domain = url.host,
+            error.code == Constants.frameLoadInterruptedErrorCode {
+            failingUrls.insert(domain)
+        }
+        
+        // wait before showing errors in case they recover automatically
+        DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
+            self.showErrorNow()
+        }
     }
-
-    func webView(_ webView: WKWebView, didReceiveLongPressForUrl url: URL, atPoint point: Point) {
-        launchLongPressMenu(atPoint: point, forUrl: url)
+    
+    public func webView(_ webView: WKWebView,
+                        decidePolicyFor navigationAction: WKNavigationAction,
+                        decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
+        
+        let decision = decidePolicyFor(navigationAction: navigationAction)
+        
+        if let url = navigationAction.request.url,
+            decision == .allow,
+            appUrls.isDuckDuckGoSearch(url: url) {
+            StatisticsLoader.shared.refreshRetentionAtb()
+        }
+        
+        decisionHandler(decision)
     }
-
-    func webView(_ webView: WKWebView, didUpdateHasOnlySecureContent hasOnlySecureContent: Bool) {
-        guard webView.url?.host == siteRating?.url.host else { return }
-        siteRating?.hasOnlySecureContent = hasOnlySecureContent
-        updateSiteRating()
+    
+    private func decidePolicyFor(navigationAction: WKNavigationAction) -> WKNavigationActionPolicy {
+        
+        if navigationAction.isTargettingMainFrame()
+            && tld.domain(navigationAction.request.mainDocumentURL?.host) != tld.domain(lastUpgradedDomain) {
+            lastUpgradedDomain = nil
+        }
+        
+        guard let url = navigationAction.request.url else {
+            return .allow
+        }
+        
+        guard !url.absoluteString.hasPrefix("x-apple-data-detectors://") else {
+            return .cancel
+        }
+        
+        guard let documentUrl = navigationAction.request.mainDocumentURL else {
+            return .allow
+        }
+        
+        if shouldReissueSearch(for: url) {
+            reissueSearchWithStatsParams(for: url)
+            return .cancel
+        }
+        
+        if !failingUrls.contains(url.host ?? ""),
+            navigationAction.isTargettingMainFrame(),
+            let upgradeUrl = httpsUpgrade.upgrade(url: url) {
+            
+            lastUpgradedDomain = upgradeUrl.host
+            load(url: upgradeUrl)
+            
+            return .cancel
+        }
+        
+        if shouldLoad(url: url, forDocument: documentUrl) {
+            return .allow
+        }
+        
+        return .cancel
     }
-
-    func webpageCanGoBackForwardChanged() {
-        delegate?.tabLoadingStateDidChange(tab: self)
+    
+    private func showErrorNow() {
+        guard let error = lastError else { return }
+        hideProgressIndicator()
+        
+        let code = (error as NSError).code
+        if  ![Constants.unsupportedUrlErrorCode, Constants.urlCouldNotBeLoaded].contains(code) {
+            showError(message: error.localizedDescription)
+        }
+        
+        webpageDidFailToLoad()
+        checkForReloadOnError()
     }
-
-    func webView(_ webView: WKWebView, didChangeUrl url: URL?) {
-        delegate?.tabLoadingStateDidChange(tab: self)
+    
+    private func isHttpsUpgradeSite(url: URL) -> Bool {
+        return url.isHttps() && HTTPSUpgrade.shared.isInUpgradeList(url: url)
     }
+    
+    private func reportHttpsUpgradeSiteError(url: URL, error: String) {
+        guard let host = url.host else { return }
+        let params = [
+            Pixel.EhdParameters.errorCode: error,
+            Pixel.EhdParameters.url: "https://\(host)"
+        ]
+        Pixel.fire(pixel: .httpsUpgradeSiteError, withAdditionalParameters: params)
+        statisticsStore.httpsUpgradesFailures += 1
+    }
+}
 
+extension TabViewController: ContentBlockerSettingsChangeDelegate {
+    func contentBlockerSettingsDidChange() {
+        onContentBlockerConfigurationChanged()
+    }
+}
+
+extension TabViewController: PrivacyProtectionDelegate {
+    func omniBarTextTapped() {
+        chromeDelegate?.omniBar.becomeFirstResponder()
+    }
+}
+
+extension TabViewController: WKUIDelegate {
+    public func webView(_ webView: WKWebView,
+                        createWebViewWith configuration: WKWebViewConfiguration,
+                        for navigationAction: WKNavigationAction,
+                        windowFeatures: WKWindowFeatures) -> WKWebView? {
+        webView.load(navigationAction.request)
+        return nil
+    }
+    
+    public func webViewWebContentProcessDidTerminate(_ webView: WKWebView) {
+        delegate?.tabContentProcessDidTerminate(tab: self)
+    }
 }
 
 extension TabViewController: UIPopoverPresentationControllerDelegate {
@@ -615,12 +1007,18 @@ extension TabViewController: UIPopoverPresentationControllerDelegate {
     }
 }
 
-extension TabViewController {
-    override func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
+extension TabViewController: UIGestureRecognizerDelegate {
+    func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
         if isShowBarsTap(gestureRecognizer) {
             return true
         }
-        return super.gestureRecognizerShouldBegin(gestureRecognizer)
+        if gestureRecognizer is WebLongPressGestureRecognizer {
+            let x = Int(gestureRecognizer.location(in: webView).x)
+            let y = Int(gestureRecognizer.location(in: webView).y)
+            let url = webView.getUrlAtPointSynchronously(x: x, y: y)
+            return url != nil
+        }
+        return false
     }
 
     private func isShowBarsTap(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
@@ -635,25 +1033,18 @@ extension TabViewController {
         return y > (view.frame.size.height - chromeDelegate.toolbarHeight)
     }
 
-    override func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer,
-                                    shouldBeRequiredToFailBy otherGestureRecognizer: UIGestureRecognizer) -> Bool {
-        if gestureRecognizer == showBarsTapGestureRecogniser {
-            return true
-        }
-        return super.gestureRecognizer(gestureRecognizer, shouldBeRequiredToFailBy: otherGestureRecognizer)
+    func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer,
+                           shouldBeRequiredToFailBy otherRecognizer: UIGestureRecognizer) -> Bool {
+        return gestureRecognizer == showBarsTapGestureRecogniser ||
+               gestureRecognizer is WebLongPressGestureRecognizer
     }
 }
 
-extension TabViewController: ContentBlockerSettingsChangeDelegate {
-    func contentBlockerSettingsDidChange() {
-        onContentBlockerConfigurationChanged()
+private class WebLongPressGestureRecognizer: UILongPressGestureRecognizer {}
+
+private extension WKNavigationAction {
+    
+    func isTargettingMainFrame() -> Bool {
+        return targetFrame?.isMainFrame ?? false
     }
-}
-
-extension TabViewController: PrivacyProtectionDelegate {
-
-    func omniBarTextTapped() {
-        chromeDelegate?.omniBar.becomeFirstResponder()
-    }
-
 }
